evd/requirements.txt
AddOns==0.7
backports.functools-lru-cache==1.6.1
BytecodeAssembler==0.6.1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Cheetah==2.4.4
cheroot==8.2.1
CherryPy==17.4.2
chronograph==1.0.0
configobj==5.0.6
contextlib2==0.6.0.post1
DecoratorTools==1.8
Django==1.4
django-mailer==0.1.0
django-notification==0.2
django-registration==0.8
django-social-auth==0.7.28
Extremes==1.1.1
feedparser==5.2.1
FormEncode==1.3.1
Genshi==0.7.3
google-api-python-client==1.7.11
google-auth==1.7.2
google-auth-httplib2==0.0.3
html5lib==1.0b8
httplib2==0.14.0
idna==2.8
jaraco.functools==2.0
Markdown==3.1.1
more-itertools==5.0.0
mysql==0.0.2
mysql-connector-python==8.0.18
MySQL-python==1.2.5
mysqlclient==1.3.4
notification==0.2.1
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.1.0
Paste==3.2.3
PasteDeploy==2.0.1
PasteScript==3.2.0
PEAK-Rules===0.5a1.dev-r2686
Pillow==6.2.1
pisa==3.0.33
portend==2.6
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyDrive==1.3.1
PyJWT==1.7.1
PyPDF2==1.26.0
python-dateutil==2.8.1
python-openid==2.2.5
python-social-auth==0.2.14
pytz==2019.3
PyYAML==5.2
registration==0.50
reportlab==3.5.32
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
simplejson==3.17.0
six==1.13.0
South==1.0.2
SymbolType==1.0
tempora==1.14.1
tgMochiKit==1.4.2
TGScheduler==1.7.0
TurboGears==1.5.1
TurboJson==1.3.2
uritemplate==3.0.0
urllib3==1.25.7
webencodings==0.5.1
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
zc.lockfile==2.0
